Moved Fragment and adapter to the package
