+++ app/src/main/java/sistemas/puc/com/finantialapp/indice/IndiceFragment.java
@@ -1,4 +1,4 @@
-package sistemas.puc.com.finantialapp;
+package sistemas.puc.com.finantialapp.indice;
 
 import android.database.Cursor;
 import android.net.Uri;
@@ -13,8 +13,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import sistemas.puc.com.finantialapp.R;
 import sistemas.puc.com.finantialapp.adapters.AbstractCursorAdapter;
-import sistemas.puc.com.finantialapp.adapters.IndiceCursorAdapter;
 import sistemas.puc.com.finantialapp.data.FinantialContract.IndiceEntry;
 import sistemas.puc.com.finantialapp.model.Database;
 import sistemas.puc.com.finantialapp.util.DividerItemDecoration;
